+++ src/storage/store.js
@@ -13,6 +13,7 @@ import { EventTarget } from "event-target-shim";
 import { fetchRandomDefaultAvatarId, generateRandomName } from "../utils/identity.js";
 import { NO_DEVICE_ID } from "../utils/media-devices-utils.js";
 import { getDefaultTheme } from "../utils/theme.js";
+import { AAModes } from "../effects";
 
 const defaultMaterialQuality = (function () {
   const MATERIAL_QUALITY_OPTIONS = ["low", "medium", "high"];
@@ -161,7 +162,7 @@ export const SCHEMA = {
         avatarVoiceLevels: { type: "object" },
         enablePostEffects: { type: "bool", default: false },
         enableBloom: { type: "bool", default: true }, // only applies if post effects are enabled
-        aaMode: { type: "number", default: 4 } // only applies if post effects are enabled
+        aaMode: { type: "string", default: AAModes.MSAA_4X } // only applies if post effects are enabled
       }
     },
